final 1.9 23/12/2021 19h44

+++ src/main/java/fr/insa/ProjetJavaBDD/ressouces/dto/AgenceCreateModel.java
@@ -9,6 +9,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AgenceCreateModel implements Serializable{
-	private int codeAgence;
+	private long codeAgence;
 	private String adresse;
 }
