+++ src/main/java/fr/insa/ProjetJavaBDD/ressouces/dto/ClientCreateModel.java
@@ -13,7 +13,7 @@ public class ClientCreateModel implements Serializable{
 	private String nom;
 	private String prenom;
 	private int age;
-	private int telephone;
+	private long telephone;
 	private String adresse;
 	private long agenceId;
 }
